cleaned up reparse file
